use require.js
I'm not totally sure it's a good idea. The plus is the code
is more organized and self contained. If you use gameserver.js
you don't have to know it needs netplayer.js for example.

But, it means you're committed to using require.js everywhere

On the plus side, require.js forces you to remove dependencies