separate cmd from cmd data
Originially it was

    send({
      cmd: 'score',
      points: 10,
    });

Now it's

    sendCmd('score', {
      points: 10,
    });

It does make the msg slighty bigger but it possibly? makes deserialization in simpler.
It also makes what's in the data part independent of it's content. On other words the
code can do this

   var cmd = msg.cmd;
   var data = msg.data;
   sendEvent(cmd, data);

Instead of

   var cmd = msg.cmd;
   var data = msg;
   sendEvent(cmd, data);

IDK, it just seems like the code shouldn't be looking inside the payload/package.
It should just be looking at the "address" as in "deliever this opaque package of data
to the event handler for cmd".

If for some reason it's better to have cmd inside that package it's easy to merge
them in the code (although that would end up modifying the object passed in)