add games.html games menu.
This makes it easier to check out the games. On one machine you
go to http://localhost:8080/games.html, pick a game, if you want
to switch games press back in the browser, pick another.

Players go to http://localhost:8080/

I worry a little that if a player is paying attention to the URL bar
they'll see the URL and go to that page and mess up stuff. See
issue in todo.md