Make server tell browser not to cache anything.
I'm not sure this is good or bad. It's good for dev because the browser is
less likely to cache your stuff? Not sure about that. I see lots of articles
on the net about changing urls from `http://foo/someasset` to
`http://foo/someasset?randomnumber` as a way to force browsers to
not cache things.

In any case most users will be just connecting for one game so
they don't need to cache anything since it will be their first and
only visit. Or so the thinking goes