Make Examples more automated.
Main menu now will automatically go to game if there is only one running

When a controller gets a disconnect message if there is only one game
running it will go to that game

To facilitate this gameserver.js automatically guesses the URL of the controller.
It takes the url of the game as in http://localhost:8080/examples/simple/gameview.html
and changes it to http://localhost:8080/examples/simple/index.html. You can override
this by passing in a `controllerUrl` option when creating the game server.

The relayserver, when receiving that URL will change 'localhost' to the local ip address

The unity examples, since they are not running in the browser, need to specifically
set a controllerUrl.