Use both ports 80 and 8080.
This means if you start the relayserver as root it will server from
both port 80 and port 8080

This means that the unity examples just work. Before they had to supply
a URL and if you wanted to run on port 80 you needed to change the
url to 8080. Now though unity can run on port 8080 and controllers
can connect to port 80 and it just works (I hope)