Tried to see if I could use CSS rotation to *force* landscape on phones
The issue is many people have their phone orientation locked to portrait.
It's annoying for them to have to unlock it to play the games. If there
was a way to either force landscape or rotate the page and just have
everything work that would be great but I ran into all kinds of issues