Attempt to fix name entry on Android.
Issue: Because of my bad css, when you go to enter your name, it
works great on iOS but on Android Chrome the controls pop up
over the name. WTF BBQ!

I tried fixing that by setting the name's z-index higher but
I guess based on child/parent relationships that didn't work.

So, next guess was "I'll hide the controls". I tried that and
it fixed the problem for Android Chrome but now iOS was broke
in that it would end up scrolling the page down and never
go back. So, after entering the name I have to scroll back
to 0,0

I hope this didn't break anything else.