Update to use SpriteManager,
Also limit max number of local players to max that can fit on scale 1 map.

Note: The game will still crash if you resize the canvas it starts if the new
size doesn't have enough space for all the players.

This is because the number of players is normally determined by how many users
connect to the game. I suppose I could fix it so AI players are added and
deleted as needed but AI players were just a hack to see lots of players
on the screen.