Fix Unity3D support so it doesn't fully deserialize until objects get to the NetPlayer
This allows a few things but the most important is you can
have more than one class associated with the same cmd name.

Before, because I auto-scanned, you couldn't do that
which had issues because you might have 2 different
scenes, that both need a 'move' cmd but they couldn't
both use 'move' as their cmdName because when I scanned
for all MessageCmdData based classes I'd find 2 and
it would fail

That's fix

The other is because of that you can also have different
ones even in the same game as they are now per NetPlayer
and not global.