add storage.js
It might be nice to abstract storage? Basically I wanted to store a 12k dataURL.
I first tried a cookie but I think cookies are limited to 4k
I then tried localstorge. It works on Chrome but not iOS Safari. It might work there
too if there's a domain? Under a IP address it's certainly not working.
I should check if it requires a domain. (not sure how I'd fix it)

But, I could also try the fileAPI? Though if safari is blocking localstorage
for ip addresses there's arguably no reason it wouldn't block the fileapi
for the same reasons