make HFT controllers use HFTSystem to find next game to switch too. This makes it much more responsive
make HFT controllers not able to connect if a game is not running. This will prevent people getting stuck on
a controller which I've seen. They'd be in a game, we'd switch games, they'd for some reason refresh or whatever
and the old controller would start and be waiting for the game. Now the controller will switch go back to "/"
which will then redirect it to whatever is running

make index.html use HFTSystem to list the games. This makes it much more responsive.