Call checkForPrerequesites
Note: I only call this if --app-mode is on. That's not actually correct
in that even if you run happyfuntimes without using --app-mode
it still needs the prerequisites. I guess my thinking was
currently they are only needed on linux and only if you install
something from superhappyfuntimes.

hmmm. maybe it should ways check?